refacto: handle message content change events in VqaViewer only
