--- ui/components/vqaCanvas/src/lib/index.ts
@@ -1,7 +1,0 @@
-/*-------------------------------------
-Copyright: CEA-LIST/DIASI/SIALV/LVA
-Author : pixano@cea.fr
-License: CECILL-C
--------------------------------------*/
-
-export * from "./utils";
