--- ui/components/vqaCanvas/src/lib/utils/index.ts
@@ -1,7 +1,0 @@
-/*-------------------------------------
-Copyright: CEA-LIST/DIASI/SIALV/LVA
-Author : pixano@cea.fr
-License: CECILL-C
--------------------------------------*/
-
-export * from "./createUpdatedMessage";
